refactor(harness): remove obsolete dependency bag
