+++ apps/api/src/harnesses/decopilot/built-in-tools/index.ts
@@ -204,9 +204,8 @@ async function buildAllTools(
     objectStorage: ctx.objectStorage,
   });
   if (userId) {
-    // Hosted `interests.write` hook: closes over ctx/storage and forwards the
-    // org/agent/user carried in the InterestsWrite payload. The tool itself no
-    // longer touches StudioContext (HarnessDeps conversion).
+    // Keep the portable tool independent of StudioContext by injecting its
+    // one storage operation here.
     tools.update_interests = createUpdateInterestsTool({
       write: async (input) => {
         await ctx.storage.interests.setForAgent(
@@ -339,9 +338,8 @@ async function buildAllTools(
   // image tier with a different credential than the chat tier (caller
   // aliases it to `provider` when they share a credential).
   if (imageProvider && models.image && ctx.objectStorage) {
-    // Cluster builds the `objectStorage` + `allowHttpExternalUrls` hooks from
-    // StudioContext + settings; the tool itself no longer reads either
-    // (HarnessDeps conversion).
+    // Resolve storage and URL policy here so the portable tool reads neither
+    // StudioContext nor settings.
     // generate_image is slow (tens of seconds). When a background dispatcher
     // is wired (cluster, hosted runs) it's made backgroundable: the call
     // enqueues a durable job and returns immediately so the turn finishes and
@@ -361,8 +359,8 @@ async function buildAllTools(
     ) as ReturnType<typeof createGenerateImageTool>;
   }
   // web_search (quick) and deep_research (deep) both consume the cluster-built
-  // `researchJob` async-gen hook (HarnessDeps conversion, spec §6). The
-  // provider/DB lifecycle lives in `createClusterResearchJob`; the tools only
+  // research job. The provider/DB lifecycle lives in
+  // `createClusterResearchJob`; the tools only
   // drive the generator. Each tier resolves its own provider so it can use a
   // different model/credential than the chat model (e.g. Gemini deep research
   // via Google while chat is on LiteLLM). Hook presence is the gate — desktop
@@ -410,10 +408,8 @@ async function buildAllTools(
   }
   // take_screenshot, scrape_url, inspect_page require Browserless API token.
   if (process.env.BROWSERLESS_TOKEN) {
-    // Cluster builds the `browserless` + `objectStorage` hooks; the tools
-    // themselves no longer read ctx or process.env (HarnessDeps conversion).
-    // The Browserless gate stays env-based — `deps.browserless` presence
-    // equals `!!process.env.BROWSERLESS_TOKEN` as set by the cluster hook.
+    // Resolve Browserless and storage here so the portable tools do not read
+    // StudioContext or environment variables.
     const browserless = {
       baseUrl: BROWSERLESS_BASE_URL,
       token: process.env.BROWSERLESS_TOKEN,
